@@ -9,28 +9,32 @@
     "@material-ui/lab": "^4.0.0-alpha.60",
     "@mdi/js": "^6.9.96",
     "@mdi/react": "^1.6.0",
-    "@nivo/core": "^0.70.0",
-    "@nivo/line": "^0.70.0",
-    "@nivo/tooltip": "^0.70.0",
-    "@react-pdf/renderer": "^1.6.17",
-    "html2canvas": "^1.3.2",
+    "@nivo/core": "^0.79.0",
+    "@nivo/line": "^0.79.1",
+    "@nivo/tooltip": "^0.79.0",
+    "@react-pdf/renderer": "^2.3.0",
+    "assert": "^2.0.0",
+    "browserify-zlib": "^0.2.0",
+    "buffer": "^6.0.3",
+    "html2canvas": "^1.4.1",
     "lodash": "^4.17.21",
-    "mobx": "^6.3.3",
-    "mobx-react-lite": "^3.2.1",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-csv": "^2.0.3",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "yup": "^0.32.9"
+    "mobx": "^6.6.1",
+    "mobx-react-lite": "^3.4.0",
+    "process": "^0.11.10",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-csv": "^2.2.2",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "stream-browserify": "^3.0.0",
+    "util": "^0.12.4",
+    "yup": "^0.32.11"
   },
   "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
-    "build": "react-scripts --openssl-legacy-provider build",
-    "eject": "react-scripts eject",
-    "test": "jest",
-    "test:watch": "jest --watch"
+    "start": "craco start",
+    "build": "craco build",
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -51,18 +55,12 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/dom": "^8.7.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.2.1",
+    "@craco/craco": "^6.4.4",
     "@types/lodash": "^4.14.182",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.10.3",
-    "@types/react": "^17.0.27",
-    "@types/react-csv": "^1.1.2",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.1",
-    "@types/yup": "^0.29.13",
+    "@types/react": "^18.0.15",
+    "@types/react-csv": "^1.1.3",
+    "@types/react-dom": "^18.0.6",
+    "@types/yup": "^0.29.14",
     "sass": "^1.53.0",
     "typescript": "^4.7.4"
   }
