@@ -9,39 +9,40 @@ import AddressForm from "./pages/AddressForm/AddressForm";
 import AnalysisAssumptionsForm from "./pages/AnalysisAssumptionsForm/AnalysisAssumptionsForm";
 import {Store} from "./ApplicationStore";
 import {observer} from "mobx-react-lite";
-import RedirectWhen from "../components/RedirectWhen";
 import Constants from "../Constants";
+import {useNavigate} from "react-router-dom";
 
 /*
  * Wrapper component for the main form application. Redirects to results if finished.
  */
 const Application = observer(() => {
+    const navigate = useNavigate();
     const store = useContext(Store);
     const [finished, setFinished] = useState(false);
 
-    return <>
-        <RedirectWhen predicate={finished} to={Constants.routes.RESULTS}/>
-        <StepperNav onFinish={() => setFinished(true)}>
-            <StepperPage label={"Address"} isDone={() => store.addressFormStore.isDone}>
-                <AddressForm/>
-            </StepperPage>
-            <StepperPage label={"Analysis Assumptions"} isDone={() => store.analysisAssumptionsFormStore.isDone}>
-                <AnalysisAssumptionsForm/>
-            </StepperPage>
-            <StepperPage label={"Electrical Rate"} isDone={() => store.electricalCostFormStore.isDone}>
-                <ElectricalRateForm/>
-            </StepperPage>
-            <StepperPage label={"Solar PV System"} isDone={() => store.solarSystemFormStore.isDone}>
-                <SolarSystemForm/>
-            </StepperPage>
-            <StepperPage label={"Solar PV Costs"} isDone={() => store.costsFormStore.isDone}>
-                <CostsForm/>
-            </StepperPage>
-            <StepperPage label={"SREC"} isDone={() => store.srecFormStore.isDone}>
-                <SrecForm/>
-            </StepperPage>
-        </StepperNav>
-    </>
+    if (finished)
+        navigate(Constants.routes.RESULTS)
+
+    return <StepperNav onFinish={() => setFinished(true)}>
+        <StepperPage label={"Address"} isDone={() => store.addressFormStore.isDone}>
+            <AddressForm/>
+        </StepperPage>
+        <StepperPage label={"Analysis Assumptions"} isDone={() => store.analysisAssumptionsFormStore.isDone}>
+            <AnalysisAssumptionsForm/>
+        </StepperPage>
+        <StepperPage label={"Electrical Rate"} isDone={() => store.electricalCostFormStore.isDone}>
+            <ElectricalRateForm/>
+        </StepperPage>
+        <StepperPage label={"Solar PV System"} isDone={() => store.solarSystemFormStore.isDone}>
+            <SolarSystemForm/>
+        </StepperPage>
+        <StepperPage label={"Solar PV Costs"} isDone={() => store.costsFormStore.isDone}>
+            <CostsForm/>
+        </StepperPage>
+        <StepperPage label={"SREC"} isDone={() => store.srecFormStore.isDone}>
+            <SrecForm/>
+        </StepperPage>
+    </StepperNav>
 });
 
 export default Application;
